% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microb_initial_guess.R
\name{initial_variational_gaussian}
\alias{initial_variational_gaussian}
\title{Gives default initial guesses for logistic-normal multinomial biclustering algorithm.}
\usage{
initial_variational_gaussian(W_count, G, Q_g, cov_str, X)
}
\arguments{
\item{W_count}{The microbiome count matrix that you want to analyze.}

\item{G}{The number of component}

\item{Q_g}{The number of biclusters for each component, a vector.}

\item{cov_str}{The covaraince structure you choose, there are 16 different models belongs to
this family:UUU, UUG, UUD, UUC, UGU, UGG, UGD, UGC, GUU, GUG, GUD, GUC, GGU, GGG, GGD, GGC.}

\item{X}{The regression covariates matrix, which generated by model.matrix.}
}
\value{
\code{new_pi_g} Initial guess of proportion

\code{new_mu_g} Initial guess of mean vector

\code{new_sig_g} Initial guess of covariance matrix for each component

\code{new_T_g} Initial guess of covariance of latent variable: u

\code{new_B_g} Initial guess of bicluster membership

\code{new_D_g} Initial guess of error matrix

\code{new_m} Initial guess of variational mean

\code{new_V} Initial guess of variational varaince

\code{new_beta_g} Initial guess of covariates coefficients.
}
\description{
Gives default initial guesses for logistic-normal multinomial biclustering algorithm.
}
