% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Micro_bi_jensens.R
\name{Mico_bi_jensens}
\alias{Mico_bi_jensens}
\title{run main microbiome bicluster algorithm.}
\usage{
Mico_bi_jensens(
  W_count,
  G,
  Q_g,
  pi_g,
  mu_g,
  sig_g,
  V,
  m,
  B_g,
  T_g,
  D_g,
  cov_str,
  iter,
  const,
  beta_g,
  X
)
}
\arguments{
\item{W_count}{The microbiome count matrix that you want to analyze.}

\item{G}{The number of component}

\item{Q_g}{The number of biclusters for each component, a vector.}

\item{pi_g}{A vector of initial guesses of component proportion}

\item{mu_g}{A list of initial guess of mean vector}

\item{sig_g}{A list of initial guess of covariance matrix for each component}

\item{V}{A list of initial guess of variational varaince}

\item{m}{A list of initial guess of variational mean}

\item{B_g}{A list of initial guess of bicluster membership}

\item{T_g}{A list of initial guess of covariance of latent variable: u}

\item{D_g}{A list of initial guess of error matrix}

\item{cov_str}{The covaraince structure you choose, there are 16 different models belongs to
this family:UUU, UUG, UUD, UUC, UGU, UGG, UGD, UGC, GUU, GUG, GUD, GUC, GGU, GGG, GGD, GGC.}

\item{iter}{Max iterations, default is 150.}

\item{const}{the permutation constant in multinomial distribution. Calculated before the main
algorithm in order to save computation time.}

\item{beta_g}{initial guess of covariates coefficients.}

\item{X}{The regression covariates matrix, which generates by model.matrix.}
}
\value{
z_ig Estimated latent variable z

cluster Component labels

mu_g Estimated component mean

pi_g Estimated component proportion

B_g Estimated bicluster membership

T_g Estimated covariance of latent variable u

D_g Estimated error covariance

COV Estimated sparsity component covariance

beta_g Estimated covariates coefficients.

sigma Estimated original component covariance

overall_loglik Complete log likelihood value for each iteration

ICL ICL value

BIC BIC value

AIC AIC value
}
\description{
run main microbiome bicluster algorithm.
}
