% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection_lasso.R
\name{model_selection_lasso}
\alias{model_selection_lasso}
\title{Model selections for \code{plnmfa}}
\usage{
model_selection_lasso(W_count, K, Q_K, model, range_tuning, iter, const, X)
}
\arguments{
\item{W_count}{The microbiome count matrix that you want to analyze.}

\item{K}{A specific number of component}

\item{Q_K}{A specific number of latent dimension.}

\item{model}{A specific model name, UUU or GUU}

\item{range_tuning}{A range of tuning parameters specified, ranged from 0-1.}

\item{iter}{Max iterations, defaul is 150.}

\item{const}{Constant permutation term in multinomial distribution.}

\item{X}{The regression covariates matrix, which generates from model.matrix.}
}
\value{
A dataframe that contain the cov_str, K, Q, AIC, BIC, ICL values for model. There may
be a lot rows if long range of tuning parameters.
}
\description{
fit several models for plnmfa along with 3 criteria values: AIC BIC and ICL
}
